Add test to demonstrate owl:Axiom interpretations in inferencing
This patch adds one test to give the RDFLib OWL-RL project's expected
behavior.

From the OWL 2 Mapping to RDF document (cited in test), it appears that
Table 7, row 7, implies a class declaration axiom should be induced by a
populated `owl:Axiom` ("populated" meaning annotated{Source,Property,
Target} are all given), whether or not the referenced axiom is directly
asserted (written as a triple).  The left column pattern contrasts with
Table 17, row 1, which specifies the triple must already be asserted (to
receive further annotations).

This patch runs up against an undefined behavior in the OWL-RDF mapping:
What should be deduced from an IRI that is an `owl:Axiom`? A subtle
detail throughout the document is `owl:Axiom` only appears in references
to blank nodes (`_:x`, rather than `*:x`).

A follow-on patch may be necessary to clarify decisions (and assertions)
around what may or may not be undefined behavior.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>